backend/pyproject.toml
[tool.poetry]
name = "adthub"
version = "0.1.0"
description = "ADT Hub V2 Backend"
authors = ["At Dawn Tech"]
packages = [{ include = "adthub", from = "src" }]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115"
uvicorn = {extras = ["standard"], version = "^0.34"}
sqlalchemy = "^2.0"
alembic = "^1.13"
psycopg2-binary = "^2.9"
pydantic-settings = "^2.0"
structlog = "^24.0"
pyjwt = "^2.8"
slowapi = "^0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-cov = "^4.0"
pytest-asyncio = "^0.23"
factory-boy = "^3.3"
faker = "^24.0"
ruff = "^0.4"
mypy = "^1.10"
bandit = {extras = ["toml"], version = "^1.7"}

[tool.bandit]
exclude_dirs = ["tests"]
skips = []

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--tb=short",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-fail-under=80",
]
markers = [
    "unit: marks tests as unit tests (no external dependencies)",
    "integration: marks tests as integration tests (requires database)",
    "schema: marks tests as schema constraint tests (requires database)",
]

[tool.coverage.run]
source = ["src"]
omit = [
    "src/*/api/*",
    "src/*/models/*",
    "src/*/config.py",
    "src/*/exceptions.py",
    "src/*/main.py",
]

[tool.coverage.report]
fail_under = 80
show_missing = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
